+++ .github/workflows/build.yml
@@ -1,0 +1,14 @@
+name: Build Catalogs
+
+on:
+  workflow_dispatch:
+
+jobs:
+  build-catalogs:
+    runs-on: ubuntu-22.04
+    steps:
+    - uses: actions/checkout@v4
+    - name: Install dependencies
+      run: make install
+    - name: Build Tycho
+      run: make tycho
