+++ src/bigsky/loaders/run.py
@@ -2,9 +2,7 @@ import sys
 
 from bigsky.loaders.utils import init_db
 from bigsky.loaders.tycho1 import load_tycho_1
-from bigsky.loaders.tycho2 import load_tycho_2
 from bigsky.loaders.ongc import load_ongc
-from bigsky.loaders.wds import load_wds
 
 
 if __name__ == "__main__":
